REFACTOR: Combine contracts and base into common
