--- src/conversion/azure/base_azure_converter.ts
@@ -1,19 +1,0 @@
-import {NodeSpec} from '../../graph';
-
-import {IEntityStore, ItemMoniker} from '..';
-
-import {AnyAzureObject} from '../azure';
-
-export function parseMonikers(item: AnyAzureObject): ItemMoniker[] {
-  return [{item, name: item.name}];
-}
-
-export function skipProcessingNodeSpecs(
-  // eslint-disable-next-line  @typescript-eslint/no-unused-vars
-  input: AnyAzureObject,
-  // eslint-disable-next-line  @typescript-eslint/no-unused-vars
-  store: IEntityStore<AnyAzureObject>
-): NodeSpec[] {
-  const empty: NodeSpec[] = [];
-  return empty;
-}
