+++ src/conversion/azure/index.ts
@@ -1,5 +1,4 @@
-export * from './contracts';
-export * from './base_azure_converter';
+export * from './common';
 export * from './converter_store';
 export * from './azure_converter';
 export * from './entity_store';
